Fix compilation issue and add the first test to fimDBHelper namespace
