+++ src/syscheckd/db/CMakeLists.txt
@@ -77,18 +77,15 @@ if(CMAKE_SYSTEM_NAME STREQUAL "Windows")
   add_library(fimdb STATIC ${CMAKE_SOURCE_DIR}/src/db.cpp
                            ${CMAKE_SOURCE_DIR}/src/file.cpp
                            ${CMAKE_SOURCE_DIR}/src/fimDB.cpp
-                           ${CMAKE_SOURCE_DIR}/src/fimDBHelper.cpp
                            ${CMAKE_SOURCE_DIR}/src/registry.cpp)
 elseif(CMAKE_SYSTEM_NAME STREQUAL "Darwin")
   add_library(fimdb STATIC ${CMAKE_SOURCE_DIR}/src/db.cpp
                            ${CMAKE_SOURCE_DIR}/src/file.cpp
-                           ${CMAKE_SOURCE_DIR}/src/fimDB.cpp
-                           ${CMAKE_SOURCE_DIR}/src/fimDBHelper.cpp)
+                           ${CMAKE_SOURCE_DIR}/src/fimDB.cpp)
 else()
   add_library(fimdb SHARED ${CMAKE_SOURCE_DIR}/src/db.cpp
                            ${CMAKE_SOURCE_DIR}/src/file.cpp
-                           ${CMAKE_SOURCE_DIR}/src/fimDB.cpp
-                           ${CMAKE_SOURCE_DIR}/src/fimDBHelper.cpp)
+                           ${CMAKE_SOURCE_DIR}/src/fimDB.cpp)
 endif(CMAKE_SYSTEM_NAME STREQUAL "Windows")
