+++ src/syscheckd/db/tests/fimDBHelper/CMakeLists.txt
@@ -7,15 +7,11 @@ set(CMAKE_CXX_FLAGS_DEBUG "-g --coverage")
 file(GLOB fimDBHelper_UNIT_TEST_SRC
     "*.cpp")
 file(GLOB FIMDB_IMP_SRC
-    "${SRC_FOLDER}/syscheckd/db/src/fimDB.cpp"
-   )
-
-file(GLOB FIMDBHELPER_SRC "${SRC_FOLDER}/syscheckd/db/src/fimDBHelper.cpp")
+    "${SRC_FOLDER}/syscheckd/db/src/fimDB.cpp")
 
 add_executable(fim_db_helper_unit_test
     ${fimDBHelper_UNIT_TEST_SRC}
-    ${FIMDB_IMP_SRC}
-    ${FIMDBHELPER_SRC})
+    ${FIMDB_IMP_SRC})
 
 if(CMAKE_SYSTEM_NAME STREQUAL "Windows")
     target_link_libraries(fim_db_helper_unit_test
