+++ src/syscheckd/db/tests/fimDBHelper/fimDBHelperTest.cpp
@@ -13,22 +13,15 @@
 #include "fimDBHelperTest.h"
 #include "syscheck.h"
 
-void FIMHelperTest::SetUp()
-{
-    
-}
+void FIMHelperTest::SetUp(){}
 
-void FIMHelperTest::TearDown()
-{
-}
+void FIMHelperTest::TearDown(){}
 
-TEST(FIMHelperTest, insert_item_to_database_success) {
+TEST_F(FIMHelperTest, insert_item_to_database_success) {
     std::string tableName = "file_entry";
     nlohmann::json insertItem;
     EXPECT_CALL(FIMDBMOCK::getInstance(), insertItem(testing::_)).WillOnce(testing::Return(0));
     int expected_return = 0;
     int return_code = FIMDBHelper::insertItem<FIMDBMOCK>(tableName, insertItem);
     ASSERT_EQ(return_code, expected_return);
-    // EXPECT_CALL(FIMDBMOCK::getInstance(), insertItem(testing::_)).WillOnce(testing::Return(0));
-    // FIMDBHelper::insertItem<FIMDBMOCK>(aux, item);
 }
